Modify 2-do_deploy_web_static.py

+++ 2-do_deploy_web_static.py
@@ -19,21 +19,36 @@ def do_deploy(archive_path):
         return False
     name_tgz = archive_path.split('/')[-1]
     name_notgz = name_tgz.split('.')[0]
-    try:
-        put(archive_path, "/tmp/")
-        sudo("rm -rf /data/web_static/releases/{}/".format(name_notgz))
-        sudo("mkdir -p /data/web_static/releases/{}/".format(name_notgz))
-        sudo("tar -xzf /tmp/{} -C /data/web_static/releases/{}/".format(
-            name_tgz, name_notgz))
-        sudo("rm /tmp/{}".format(name_tgz))
-        sudo(("mv /data/web_static/releases/{}/web_static/* " +
-              "/data/web_static/releases/{}/").format(
-                  name_notgz, name_notgz))
-        sudo("rm -rf /data/web_static/releases/{}/web_static".format(
-             name_notgz))
-        sudo("rm -rf /data/web_static/current")
-        sudo("ln -s /data/web_static/releases/{}/ /data/web_static/current"
-             .format(name_notgz))
-        return True
-    except Exception:
+    task = put(archive_path, "/tmp/")
+    if task.failed is True:
         return False
+    task = sudo("rm -rf /data/web_static/releases/{}/".format(name_notgz))
+    if task.failed is True:
+        return False
+    task = sudo("mkdir -p /data/web_static/releases/{}/".format(name_notgz))
+    if task.failed is True:
+        return False
+    task = sudo("tar -xzf /tmp/{} -C /data/web_static/releases/{}/".format(
+        name_tgz, name_notgz))
+    if task.failed is True:
+        return False
+    task = sudo("rm /tmp/{}".format(name_tgz))
+    if task.failed is True:
+        return False
+    task = sudo(("mv /data/web_static/releases/{}/web_static/* " +
+                 "/data/web_static/releases/{}/").format(
+                     name_notgz, name_notgz))
+    if task.failed is True:
+        return False
+    task = sudo("rm -rf /data/web_static/releases/{}/web_static".format(
+        name_notgz))
+    if task.failed is True:
+        return False
+    task = sudo("rm -rf /data/web_static/current")
+    if task.failed is True:
+        return False
+    task = sudo("ln -s /data/web_static/releases/{}/ /data/web_static/current"
+                .format(name_notgz))
+    if task.failed is True:
+        return False
+    return True
